image-consumer/requirements.txt
torch==2.0.0
pillow==9.0.1
requests==2.29.0
transformers==4.24.0
mlflow==2.3.0
pandas==1.5.3
pyspark==3.4.0
